backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
transformers>=4.37.0
torch>=2.4.0
pydantic>=2.6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
numpy>=1.26.0
pandas>=2.0.0
